requirements.txt
Flask==2.2.3
Flask_JWT_Extended==4.4.4
Flask_Login==0.6.2
Flask_Mail==0.9.1
flask_sqlalchemy==3.0.3
itsdangerous==2.1.2
PyJWT==2.6.0
python-dotenv==1.0.0
Werkzeug==2.2.3
